pyproject.toml
[project]
name = "realtime_behavior_intelligence"
version = "0.1.0"
description = "Tracks apps and network usage and applies ML to detect anomaly from normal behavior"
authors = [
    {name = "Lakshya Saharan",email = "lakshyasaharan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "psutil (>=7.0.0,<8.0.0)",
    "scapy (>=2.6.1,<3.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "influxdb3-python (>=0.14.0,<0.15.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "tensorflow (>=2.19.0,<3.0.0)",
    "kafka-python (>=2.2.15,<3.0.0)",
]

[tool.poetry]
packages = [{include = "src"}]  # This tells Poetry where your Python packages are

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
